Move _theme.scss to `/UI` folder in the new `~/src/components`
Also moves `Cockpit FC`

Addresses #6 (Create `_theme.scss` for misc globals)
Project: 'Import my standard CRA libraries'
