+++ src/App.js
@@ -1,6 +1,6 @@
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { BrowserRouter, Route } from 'react-router-dom';
-import Cockpit from './Cockpit/Cockpit';
+import Cockpit from './components/Cockpit/Cockpit';
 
 function App() {
 	return (
