+++ src/components/Cockpit/Cockpit.scss
